Revert "Merge branch 'Skytils:1.x' into nacho-order-nacho-auction"
This reverts commit c354f75ca2506bea155aa0e9ff2af725f4eb2492, reversing
changes made to bba228482198b98acdf2302af90252255f0b0a94.

+++ src/main/kotlin/gg/skytils/skytilsmod/features/impl/mining/MiningFeatures.kt
@@ -117,7 +117,7 @@ object MiningFeatures {
 
     @SubscribeEvent(priority = EventPriority.HIGHEST, receiveCanceled = true)
     fun onChat(event: ClientChatReceivedEvent) {
-        if (!Utils.inSkyblock || event.type == 2.toByte()) return
+        if (!Utils.inSkyblock || event.type.toInt() == 2) return
         val formatted = event.message.formattedText
         val unformatted = event.message.unformattedText.stripControlCodes()
         if (Skytils.config.powerGhastPing) {
